% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_babynames.R
\name{map_babynames}
\alias{map_babynames}
\title{Adding a color to the background.}
\usage{
map_babynames(name, start_year = 1910, stop_year = 2016, fancy = FALSE)
}
\arguments{
\item{name, }{name to generate a plot for}

\item{start_year, }{first year to include names. must be between 1910 and 2016}

\item{stop_year, }{last year to include names. must be between 1910 and 2016}

\item{fancy, }{whether or not to change the plot font to Raleway and use theme_fancy() to add background color. needs the google Raleway font downloaded to work (https://fonts.google.com/specimen/Raleway). Defaults to false.}
}
\description{
Adding a color to the background.
}
